logging: frontend: stmesp_demux: Improve robustness
Demultiplexer was not ready to handle case when log message was
incomplete which was followed by other log messages. Such scenario
could occur if there was a fault that happen during logging of
a message. In that case incomplete message was followed by valid
messages (fault report) and this fault report was not handled
because processing was blocked waiting for completion of a
message which preceeded fault report.

Since it is expected that some messages may be incomplete a
garbage collection mechanism is added. When start of a message is
received timestamp is logged and list of incomplete messages
is checked for 'old' messages which persist in incomplete state
for long. When message timeouts it is closed and marked as
invalid. It unblocks processing of following messages.

(cherry picked from commit af391b8b3f991cb54af35569759fed60653f9463)

Original-Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>
GitOrigin-RevId: af391b8b3f991cb54af35569759fed60653f9463
Cr-Build-Id: 8736011314115091969
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8736011314115091969
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ic3f4ca1e4851dc1396f43b5c233cf5233f7aff29
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5881488
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>