drivers: wifi: Fix mutex re-initialization
The mutex is used to protect RPU zephyr context which gets modified for
every interface down and up (including recovery), so, it was being
re-initialized but also used to protect down and up which is a bug.

Move the re-initialization to the driver entry so that it happens only
once and we can properly use the mutext for down and up protection.

(cherry picked from commit d30161cd8bd182de5c725376c98b19bce686c6ac)

Original-Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>
GitOrigin-RevId: d30161cd8bd182de5c725376c98b19bce686c6ac
Cr-Build-Id: 8733565196572304993
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8733565196572304993
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I56e7ab9d78c025dfc451c73b97ca8037f3658627
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5940631
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>
Commit-Queue: Eric Yilun Lin <yllin@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>