drivers: audio: dmic_nrfx_pdm: change log level in PDM read
Currently the dmic_nrfx_pdm read implementation treats a timeout as an
error, in that a logging error message is produced when no PDM data is
available. However, for non-(or minimally-)blocking applications this is
normal behavior. Classified this way, the logger is flooded with error
messages unnecessarily, unless the log level is changed for the dmic
module, which is not desirable.

This modification simply changes the log-level to debug so the
application user can then decide whether or not a failed read needs to
produce an error message.

(cherry picked from commit a1474a98685e8b6d393c25cfbed4522205c8263f)

Original-Signed-off-by: Anthony Wertz <awertz@pm.me>
GitOrigin-RevId: a1474a98685e8b6d393c25cfbed4522205c8263f
Cr-Build-Id: 8730756689262017889
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8730756689262017889
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ib793421e3783f09a556ec07da90b8cc76295b464
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6037795
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>