tests: latency_measure: reduce the chance of cycles underflow
Sometimes there's an unusually large cycles for tests that are
known to complete with just a few cycles. Upon some testing,
I found that it was because the overhead cycles was larger
than the cycles taken by tests, causing the cycles to
underflow.

To workaround that, make sure that the overhead measurement
thread runs uninterrupted, repeat the measurement for a few
times, and take the minimum value.

(cherry picked from commit 59e41ef830d49816d4a3900215a0fb55da1228d0)

Original-Signed-off-by: Yong Cong Sin <ycsin@meta.com>
Original-Signed-off-by: Yong Cong Sin <yongcong.sin@gmail.com>
GitOrigin-RevId: 59e41ef830d49816d4a3900215a0fb55da1228d0
Cr-Build-Id: 8734380568354140561
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8734380568354140561
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I88b4a05d540598b2ce5d16d060f2407fd7a58178
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5920154
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>