tests: i2c_emul: drop the array designators
Drop the array dasignators as they generate a

warning when compiling with clang. Not too sure if this is an actual
problem or if the warning should be disabled instead but since it's just
a test it should be fine to just drop them.

(cherry picked from commit e428a52a282813d8cd4cbfe376dd6aaf6edf6ed4)

Original-warning: array designators are a C99 extension
Original-Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
GitOrigin-RevId: e428a52a282813d8cd4cbfe376dd6aaf6edf6ed4
Cr-Build-Id: 8735920717536838961
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8735920717536838961
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I00e01ac3729053ef1ca7e38a5a05a3f4eac2a7cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5886552
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>