drivers: gpio: stm32: keep track of clock enabling per pin
The driver enables the clock of a gpio-port if any of the
pins use the port. This is done by calling pm_device_runtime_get
when a pin is used and pm_device_runtime_put when the pin is not
used anymore.
These calls needs to be balanced. But if a single pin was configured
as GPIO_DISCONNECTED multiple times, every time pm_device_runtime_put
was called.
This caused the clock of the port to be stopped and therefore also
other pins on the same port stopped working.

This commit fixes this by keeping track of which pin on a port
has requested the clock and only call pm_device_runtime_get or
pm_device_runtime_put when the clock-request for the specific pin
changes.

Fixes #77698

(cherry picked from commit cf837dd371f26472a1b4d2cdf628780184dc05ac)

Original-Signed-off-by: Jeroen Broersen <jbroersen@interact.nl>
GitOrigin-RevId: cf837dd371f26472a1b4d2cdf628780184dc05ac
Cr-Build-Id: 8735920717536838961
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8735920717536838961
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ibde92d0dabefb53ee54e35255c10f691af0a108e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5886527
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>