drivers: spi_mcux_lpspi: Consolidate IRQ handle
Small change to consolidate the amount of lines of code in
the spi_mcux_isr by making a macro for the argument and removing
a line of code that was commented out.

Also move the ISR to be the first function in the file which is
common in many other drivers, instead of randomly in the middle of the
file. And move the isr callback to be next to the isr.

(cherry picked from commit 595e16caf6d53f0a05e0e5013f3d0680b2ba6a75)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: 595e16caf6d53f0a05e0e5013f3d0680b2ba6a75
Cr-Build-Id: 8730031914779849585
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8730031914779849585
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I76c6d52b62512be5ec6b6e33f55754c0fb476382
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6057869
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>