Bluetooth: CAP: Add support for handling ASE errors
If we get an error/rejection from the CAP acceptor when
performing the Unicast Audio Start or Stop procedure then
we need to abort the procedure and let the application determine
what the next step is.

This change triggered a corner case when connecting to multiple
CAP acceptors as the CAP initiatior. This was also fixed as part
of this.

(cherry picked from commit 2dc1113a9487b6bd1ef64f9ef47be0b99238658a)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 2dc1113a9487b6bd1ef64f9ef47be0b99238658a
Cr-Build-Id: 8732659221460854369
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8732659221460854369
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I7b10af80cf5e392c2a5c4f6969587da4791f932c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5976013
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>