modules: Add placeholder for CONFIG_RUST
In order to be able to ifdef based on `CONFIG_RUST`, the config must be
defined in the main repo (because the rust module is optional).  Add
this as a placeholder.  If the module is not availabe, the symbol will
be present, and always disabled.  The module will provide a full
definition for the Kconfig.

(cherry picked from commit bd46e8d8496e7ba6b1e0817fc93d638932e52187)

Original-Signed-off-by: David Brown <david.brown@linaro.org>
GitOrigin-RevId: bd46e8d8496e7ba6b1e0817fc93d638932e52187
Cr-Build-Id: 8735619482632649473
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8735619482632649473
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I51a0840df74c8d09109c538a09ee0b446a443fe1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5895359
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>