Bluetooth: BAP: Disallow bt_bap_stream_stop when CIS is not connected
When the CIS for a BAP stream is not connected, we cannot truly
be in the disabling state (the only state the stop operation
can be performed by the unicast client).

The reason for this is that if the CIS is disconnected, then
the ASCS server shall transition to the QoS Configured state
regardless of whether it has received a receiver stop ready
command from the unicast client.

(cherry picked from commit 756f5f940ccb06db38fd14d738fcc98efe96d0c7)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 756f5f940ccb06db38fd14d738fcc98efe96d0c7
Cr-Build-Id: 8732659221460854369
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8732659221460854369
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Id04f0be6f3013bff98aab33de71b6e1d0f2e5185
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5976012
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>