tests: drivers: uart: uart_pm: Disable PM_DEVICE_SYSTEM_MANAGED
When PM_DEVICE_SYSTEM_MANAGED is enabled (and that's the default) all
devices are suspended when device go to idle state (through pm
subsystem and not through just cpu_idle). In this test we are manually
controlling PM of the DUT so we don't want that.

Additionally, suspend action is taking a semaphore and it is illegal
in idle thread (where system managment operates) which causes assert
in kernel scheduler.

Removed for nrf54h20 as it applies only there.

(cherry picked from commit cbe0abfd0002c761591af0d9ba7bc321d00dc589)

Original-Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>
GitOrigin-RevId: cbe0abfd0002c761591af0d9ba7bc321d00dc589
Cr-Build-Id: 8735014746731713681
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8735014746731713681
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I0c926feccbdd250abf0fbc4f8188f2f86c2e4bd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5909496
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>