drivers: intc: plic: fix for SMP=n when MP_MAX_NUM_CPUS > 1
The functions to obtain the address are hardcoded to return
the address of the first core when `CONFIG_SMP != y`, this
causes an issue with enabling an IRQ when there are more than
one core in the system (`CONFIG_MP_MAX_NUM_CPUS > 1`), as the
driver would first enable the IRQ on the first core, and when
it tries to obtain the address for the following cores and
disable the IRQ on them, the functions continue to return the
address of the first core, causing the IRQ to be disabled
on the first core.

Fix this by determine if `CONFIG_MP_MAX_NUM_CPUS > 1` instead
of `CONFIG_SMP=y` when returning the address.

(cherry picked from commit 796b795cda29441e8688702d028997689a377f40)

Original-Signed-off-by: Yong Cong Sin <ycsin@meta.com>
Original-Signed-off-by: Yong Cong Sin <yongcong.sin@gmail.com>
GitOrigin-RevId: 796b795cda29441e8688702d028997689a377f40
Cr-Build-Id: 8735014746731713681
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8735014746731713681
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: If07369d1700646df4470d411f35e40448b74f129
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5909502
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>