usb: device_next: cdc_acm: Prevent polling for buffer in TX
In case the host doesn't pull the new data from the endpoint, the work task
would schedule itself again delayed (at the max. priority). When there is
no terminal program or active application reading the endpoint this
results in a constant polling of the endpoint burning up to 5% of the
CPU cycles.

By using a atomic flag for tx busy, the polling is solved and changed into
a postponed execution of the next work task which saves up to 5% of
CPU cycles and allows a better real-time behavior for other tasks.

Secondly, if the TX interrupt is disabled but there is still data in the TX
FIFO (ring buffer), the implementation will continue to trigger subsequent
TX work and attempt to flush the data to the host.

(cherry picked from commit 5cbd1f6c39002410a814700cb75dcce189bb5295)

Original-Signed-off-by: Vincent van der Locht <vincent@synchronicit.nl>
GitOrigin-RevId: 5cbd1f6c39002410a814700cb75dcce189bb5295
Cr-Build-Id: 8732659221460854369
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8732659221460854369
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I5a10eaa47095eed2cede65996a3d999544bcc0d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5975952
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>