net: lib: sntp_simple: Handle responses from previous iterations
SNTP simple runs request iterations with exponential backoff.
If the net interface is a slower connection (ie. CAT M1 modems)
then the request will be sent but the response may take time to
be received, thus causing a timeout and another request to be sent.
Because of the nature of UDP and the fact that the same socket
(source IP/port combo) is being used for both requests, a delayed
response to the first request can be received as the response to the
second request, causing -EINVAL to be returned when the timestamps
mismatch (see subsys/net/lib/sntp/sntp.c). The solution provided
retries receiving the response when the timestamp is mismatched
(without sending an additional request).

(cherry picked from commit 72534b83e47da28862cb7696158e6503710f5e03)

Original-Signed-off-by: Marcus Penate <marcus.penate@ellenbytech.com>
GitOrigin-RevId: 72534b83e47da28862cb7696158e6503710f5e03
Cr-Build-Id: 8731572062094510337
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8731572062094510337
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I534691e8418f804b4b5837b2cb30e1c2a34d0879
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6008673
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>