bluetooth: host: LE CS subevent result reassembly
Adds HCI support for LE CS subevent result continue
event and the reassembly logic for the partial results.

When subevent results are completed or the subevent is
aborted, the user callback is invoked with a buffer
pointing to the HCI event buffer, so no copy is done.

When subevent results are incomplete, then a reassembly
buffer is allocated from a fixed sized pool. This buffer
is used for the reassembling of the subevent result
containing all of the step data, which is then passed
to the user via the callback. kconfigs have been added
to set the size and the count of the reassembly buffer.

(cherry picked from commit c5a126cedb2f96239662352ae1662424d261c53e)

Original-Signed-off-by: Burak Gorduk <burak.gorduk@nordicsemi.no>
GitOrigin-RevId: c5a126cedb2f96239662352ae1662424d261c53e
Cr-Build-Id: 8733565196572304993
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8733565196572304993
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I49bf984f616fe4d7ed1246d4b5a2503858077726
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5937718
Commit-Queue: Eric Yilun Lin <yllin@google.com>
Tested-by: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>