soc: nxp: imxrt: correct flexspi XIP check to avoid reclocking
RT11xx SOC init should check to see if the zephyr flash node is
set to a device on the FLEXSPI bus to determine if the part is running
in XIP mode. This check was incorrect, so the FLEXSPI was being
reclocked in XIP mode to 24 MHz. Fix this check so the FlexSPI is not
downclocked.

Fixes #75702

(cherry picked from commit 0856ceed7b2248c11a52b012e76b9fe6007fef8d)

Original-Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>
GitOrigin-RevId: 0856ceed7b2248c11a52b012e76b9fe6007fef8d
Cr-Build-Id: 8731934450537830641
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8731934450537830641
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I97d5bcc8c79c5e83161d340d3f090d5a73b7cede
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5999687
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>