boards: opta: ADC support
A valid device tree configuration is provided for the ADCs of the 8 input
channels and the sample adc_dt works out of the box. Obviously this is
not the only possible configuration but it provides a good template for
further customization without the need to lookup the ADC GPIOs and
connections in the schematics.

(cherry picked from commit 1f5a1b50fa7d7694b8ad6dfa6e447408e19cb521)

Original-Signed-off-by: Federico Di Gregorio <fog@dndg.it>
GitOrigin-RevId: 1f5a1b50fa7d7694b8ad6dfa6e447408e19cb521
Cr-Build-Id: 8730666093191416657
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8730666093191416657
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I4657a8954d7c62d17f621e53c3d633c15096bfc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6041400
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Tristan Honscheid <honscheid@google.com>