kernel: Begin abstracting out _sched_spinlock
In a uniprocessor system, _sched_spinlock may not need to be
held in all the same cases that it does in a multiprocessor
system. Removing those unnecessary usages can lead to better
performance on UP systems. In the case of uncontested taking
and giving of a semaphore, this can be as much as a +14%
performance gain.

(cherry picked from commit cedd36106b3fdb2159ba1e03e05daeee1a31389a)

Original-Signed-off-by: Peter Mitsis <peter.mitsis@intel.com>
GitOrigin-RevId: cedd36106b3fdb2159ba1e03e05daeee1a31389a
Cr-Build-Id: 8732749822481685601
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8732749822481685601
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I421865d77e597ea9f00b212fac4d17a0d287152f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5975784
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>