drivers: wifi: eswifi: fix casts for 64 bit pointers
The eswifi driver uses an integer to identify the socket used for each
connection, which must be cast to and from a `void *`. This causes
warnings on 64 bit platforms, as precision is lost when casting from a
`void *` to `int`. Use a `intptr_t` type to store the socket value to
resolve this warning. Also, fix a function signature for the `accept_cb`
to use `size_t` instead of `int` for the length

Fixes #80242

(cherry picked from commit 3d4f83aaace5f87c3fdd4af2611ea9bbf3da1560)

Original-Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>
GitOrigin-RevId: 3d4f83aaace5f87c3fdd4af2611ea9bbf3da1560
Cr-Build-Id: 8732659221460854369
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8732659221460854369
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I107dbc26187829356733befa44030b5151b8f51e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5978582
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>