drivers: udc_nrf: fix enqueue of control IN transfer with length 0
If the direction of the last setup packet is not to the device but to
the host, then the transfer is not a status stage and should be queued.
This is not checked and prevents a zero length control IN transfer to
the host, e.g. used by the DFU class to indicate the end of the upload
process.

(cherry picked from commit 973f914b90a2e204165e9e32cdf4a653acc00cd9)

Original-Signed-off-by: Johann Fischer <johann.fischer@nordicsemi.no>
GitOrigin-RevId: 973f914b90a2e204165e9e32cdf4a653acc00cd9
Cr-Build-Id: 8731934450537830641
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8731934450537830641
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ia99ba2d93ce55947075da72b724cff1117e6a145
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5999676
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>