pm: policy: event: use uptime ticks
Update events to use uptime ticks, which is a monotonic clock which
in the same res as kernel ticks. This makes comparisons simple and
removes the complexity of dealing with wrapping counter values.

The wrapping is particularly problematic for events since this makes
it quite complex to track if an event has occured in the past, or
will occur in the future. This info is needed to know if an event
has actually been handled or not.

(cherry picked from commit 59779ebe4b12d08bee7e8d15028fb0ee2152b4e3)

Original-Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>
GitOrigin-RevId: 59779ebe4b12d08bee7e8d15028fb0ee2152b4e3
Cr-Build-Id: 8729035348064486593
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8729035348064486593
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I1a93dd6b4d12da1d7f98cba2f415b7ea549ec0eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6079801
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>