tests: Bluetooth: Audio: Remove uses of K_FOREVER in syswg for TX
Several tests were using K_FOREVER when allocating the
buffer for TX in the system workqueue, which is illegal behavior.

The solution chosen was to create a TX thread to handle TX,
similar to the solution used in the audio shell and some
sample applications.

This way we can continue to use K_FOREVER when allocting buffers
and it will always be done in a round-robin fashion while
TXing as much as possible, by always enqueuing all the buffers
with mock data.

Since this works for all streams (both broadcast and unicast),
it was obvious to use the same implementation for all tests,
and thus cleaning up the tests a bit and more them more similar.

(cherry picked from commit 778a9afae35c91a0a4d17fe459091268a1860208)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 778a9afae35c91a0a4d17fe459091268a1860208
Cr-Build-Id: 8728854153376355025
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8728854153376355025
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I7dfce92b72526727d3b9c0c6d1f56ef0ada17261
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6087887
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>