drivers: serial: cdns: fix interrupt driven uart
The tx fifo empty interrupt is a edge driven interrupt, so if it
is already empty then and the interrupt is enabled, it will not
fire so the isr needs to be triggered manually for the callback.

This also removes the unnecessary interrupt locking in the isr and
removes the receiver timeout interrupt.

(cherry picked from commit 36772a88898d8ed75e91374ab2651c7678968b20)

Original-Signed-off-by: Ryan McClelland <ryanmcclelland@meta.com>
GitOrigin-RevId: 36772a88898d8ed75e91374ab2651c7678968b20
Cr-Build-Id: 8730937883239782945
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8730937883239782945
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I5c254e5d5ac97b2f240b816a34c0d189efae723f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6028824
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>