Bluetooth: CSIP: Handle disconnects while in procedure
If a device disconnects while we are in a procedure then
get_next_active_instance would return a service instance pointer
with the `conn` set to NULL.

The issue was caused by the set_info being potentially memset
when the device that disconnected was the one that held the
set_info pointer.
The solution is to not use a pointer, but rather a copy of the
set_info, so that the active.set_info value is still valid
after a disconnect.

Since the set_info is not longer a pointer to a specific
set_info from one of the members, the logs have been updated
as well, as the pointer of the active.set_info is useless
for debugging.

(cherry picked from commit 1f1e4afa4f145699677f8ee50dac5c5890d42dd9)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 1f1e4afa4f145699677f8ee50dac5c5890d42dd9
Cr-Build-Id: 8731934450537830641
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8731934450537830641
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I781365f9135ca89bea7aa1ac8a0a771afa472198
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6002920
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>