boards: beaglev_fire: fix memory address
Per the docs, the memory at address 0x80000000 ends at 0xC0000000.
In other words, the address space is 0x40000000, which is only half
of the size we want to map. This means that the upper address space
previously mapped was overlapping with the space reserved for non-cached
memory.

Instead, we map the entire 2GB at 0x1000000000, which is the correct
address for cached DDR that occupies more than 1 GB.

We defined a new node in the device tree for this memory region,
`beaglev.ddr_cached_high`. We did not reuse the `soc` node because
we needed to redefine the `#address-cells` to be 2, and doing so
would have affected other nodes under `soc`.

(cherry picked from commit efc2cfe326232140ed7bf0d5b9415d0b0d8d3af7)

Original-Signed-off-by: Alex Charlton <alex.n.charlton@gmail.com>
GitOrigin-RevId: efc2cfe326232140ed7bf0d5b9415d0b0d8d3af7
Cr-Build-Id: 8730122511391905569
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8730122511391905569
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I575d0ff9cd69b55ae78d23d3b50560d55c62a0af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6054153
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>