dts: bindings: base: Undeprecate label
Label property is described in DT spec and does not need to be
deprecated in base.yaml anymore. It was originally deprecated to
discourage what was previously the most common use case of labels in
zephyr which was the old device_get_binding, which was rightfully
removed. However, labels do have a purpose as described in DT spec of
providing a human readable string to software to describe the device,
which there is some use for.

The description of a label should be given in the device binding, as
stated in DT spec.

Label properties should be of type string.

(cherry picked from commit 8e87d5547371a48d420d5d4a423a8ceda7788984)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: 8e87d5547371a48d420d5d4a423a8ceda7788984
Cr-Build-Id: 8730164773885449249
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8730164773885449249
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I502e484ad106badea9acfbdee859d1370a78caf7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6053110
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>