drivers: fpga: fix waveform for iCE40 configuration in SPI mode
The datasheet of the iCE40 specifies that there should be a leading and
trailing clocks phase during its configuration with SPI. Due to the
limitations of the SPI interface, and probably also due to a lock of
support for such a feature for instance in the STM32 SPI peripheral,
this is achieved with additional SPI transfers before and after the
actual image. Unfortunately, this by default also affects the slave
select GPIO, which has to stay high during these phases.
This fixes this behaviour via not passing the slave select GPIO
to the SPI driver and manipulating this GPIO manually.

(cherry picked from commit b0a1dddde34b922032e5ccf84a2245c8c2177828)

Original-Signed-off-by: Benedikt Schmidt <benedikt.schmidt@embedded-solutions.at>
GitOrigin-RevId: b0a1dddde34b922032e5ccf84a2245c8c2177828
Cr-Build-Id: 8730286630880259329
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8730286630880259329
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Id8cc41a0d1318fd38a6875621494bbd692f60324
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6043664
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>